refactor(types): split pdf.ts god-file into cohesive type modules
The 972-line src/types/pdf.ts declaration-only god-file is split into
19 cohesive modules under src/types/pdf/ (geometry, tables, content,
document-structure, safety, layout, document-map, document-ast,
trust-report, text-layer, accessibility, result, render, regions, ocr,
region-analysis, search, source, inspection). src/types/pdf.ts becomes
a barrel that re-exports every module, so the public type surface is
unchanged and all ~30 importers of '../types/pdf.js' keep working
verbatim.

Pure mechanical move: zero logic change, every type definition copied
verbatim, cross-module references use type-only imports. Symbol set is
identical (113 exported interfaces/types, verified by diff). typecheck,
lint, full test suite, and build all green and unchanged vs baseline
(236 pass / 7 skip / 0 fail).
